requirements.txt
Flask>=2.1.1
requests==2.27.1
Flask-Limiter==2.4.1
argon2-cffi==21.3.0
python-dotenv==1.0.0
pytest==7.4.0
pytest-flask==1.2.0
cryptography==41.0.4
pyjwt==2.8.0
werkzeug==2.2.3